Extract partner data file upload into a dedicated package.
Move Batch submission, sheet parsing, Google Sheets I/O, and processing logic out of the management command and model into partner_data_file_upload/, and expand admin list search by contributor and user email.
